Renames map to Select. Adds index.
