+++ index.js
@@ -1,0 +1,7 @@
+const Select = require('./src/select');
+
+const bindAll = function() {
+  Array.prototype.Select = Select;
+};
+
+module.exports = bindAll;
